+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Added `is_html()` and `html()` to `Mantle\Http_Client\Response` to check if the
+  response is HTML and to return an `HTML` instance for making assertions
+  against the HTML content.
+
 ## v1.9.4
 
 ### Added