+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v1.9.6
+
+### Fixed
+
+- Fixed issue with `feed()` method in `Mantle\Http_Client\Response` to account for WordPress 6.5 - 6.6;
+
 ## v1.9.5
 
 ### Added